requirements.txt
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
backports.functools-lru-cache==1.5
beautifulsoup4==4.7.1
blinker==1.4
bs4==0.0.1
bzr==2.8.0.dev1
click==7.0.dev0
colorama==0.3.7
configobj==5.0.6
configparser==3.5.0b2
constantly==15.1.0
cryptography==2.3
distro==1.3.0
entrypoints==0.2.3.post3
enum34==1.1.6
folder-color==0.0.86
folder-color-common==0.0.86
futures==3.2.0
get==2019.3.22
httplib2==0.11.3
hyperlink==17.3.1
idna==2.6
incremental==16.10.1
ipaddress==1.0.17
keyring==15.1.0
keyrings.alt==3.1
launchpadlib==1.10.6
lazr.restfulclient==0.14.0
lazr.uri==1.0.3
Morfessor==2.0.4
numpy==1.16.2
oauthlib==2.0.6
PAM==0.4.2
Parsley==1.2
polyglot==16.7.4
post==2019.3.22
psutil==5.4.6
public==2019.3.22
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycld2==0.31
pycrypto==2.6.1
PyGObject==3.30.1
PyICU==2.2
PyJWT==1.6.4
pyliblzma==0.5.3
pyOpenSSL==18.0.0
pyserial==3.4
pyxdg==0.25
query-string==2019.3.22
request==2019.3.22
SecretStorage==2.3.1
service-identity==16.0.0
simplejson==3.15.0
six==1.12.0
soupsieve==1.9
torbrowser-launcher==0.2.9
Twisted==18.7.0
txsocksx==0.0.0
virtualenv==16.4.3
wadllib==1.3.3
zope.interface==4.3.2
